HADOOP-18679. My changes based on iceberg PoC
We are going to need a default FS impl which just
invokes delete(path, false) and maps any IOE to a failure.
